# MAINT: Improve performance of np.copyto(where=scalar)