Setup:
```python
x = np.arange(100000)
y = x.copy()
kw = ...

%timeit np.copyto(x, y, **kw)
```

| `kw`|  `dict()` | `dict(where=True)` | `dict(where=False)` |
|----|----|---|----|
| Before | 31 µs ± 144 ns | 222 µs ± 9.09 µs | 66.9 µs ± 2.47 µs
| After| 31 µs ± 261 ns | 33.3 µs ± 888 ns | 1.09 µs ± 7.42 ns

This speeds up `np.copyto(x, y, where=ma_array.mask)` in the case when `mask == nomask`, without the optimization having to be made at every call site.

In theory we could check for this in `array_where` before we even do an array conversion, but it's probably not worth the effort, and doesn't give an speedup in  `O(dst.size)`