@seberg 2017-07-26
I am somewhat OK with this, but frankly do not quite see the point. Do you have an actual use case?