@eric-wieser 2017-07-26
Makes passing the documented default argument not incur a substantial penalty, and also speeds up masked array code that didn't already have the optimization.