@seberg 2017-07-26
Still don't immediatly understand. The documented default is `None`, no?  As to masked array code, why would it end up with a scalar? It might end up with nomask, which is special anyway?