@eric-wieser 2017-07-26
`nomask` is a scalar.

> The documented default is None

Yeah, this turned out to be the opposite of what the default was...