@seberg 2017-08-06
Well, I will leave someone else to decide on this to be honest. I see that it might be nicer for masked arrays, so I am fine with it. I am not sure I like the default change, because I wonder if `None` as in "do not worry about this" isn't easier to understand for someone who does not really know broadcasting.