@eric-wieser 2017-08-06
@seberg, the default has not changed. Passing `None` explicitly in master is not the same as passing nothing at all - try it!