@seberg 2017-08-06
What I am saying is, that the second commit, I think I would prefer making the documentation correct rather then changing it, but that is just my gut feeling.