@seberg 2017-08-06
Of course you could argue that it might break someones wonky code, so….