@seberg 2017-08-06
So I am OK with the optimization. @charris what do you think, should we rather see it as a bug that `where=None` is not the same as not passing in `where` or avoid it by changing the documentation in a manner that hides this by setting it to `where=True` (and with this that is just the same speed anyway).