@mhvk 2017-08-06
FWIW, I think we're much better off changing the documentation to match reality, as done here.