@seberg 2017-08-07
Hehe, fine with me me, I just lean towards the other direction. I still kinda think `None` to mean "not used" as the docs suggest is nice, though it could create a bug out there somewhere in principle anyway.