@njsmith 2017-08-07
I agree that having `None` be interpreted as an alias for `False` is pretty confusing. (Why is it even doing that? Please tell me it's not accepting arbitrary objects and checking their truthiness.) But that seems like an independent issue we could deal with in its own issue/pr, and in the mean time this seems like a simple improvement over the status quo.