@eric-wieser 2017-08-07
> Please tell me it's not accepting arbitrary objects and checking their truthiness.

    >>> a = np.array(1)
    >>> np.copyto(a, 2, where=object())
    >>> a
    array(2)

Sorry