@ahaldane 2017-09-17
Re: truthiness

The truthiness conversion happens anytime we convert to a boolean array, and wheremask is documented to be an `array_like of bool`:
```python
>>> np.array([{}, (), 0, None], dtype='bool')
array([False, False, False, False], dtype=bool)
```
So I'm not sure there is anything to fix wrt to truthiness conversion.