@ahaldane 2017-09-17
As to this PR, LGTM. In a better world I would prefer to modify `array_copyto` to make `where=None` equivalent to `wheremask = NULL`, but this seems fine too and is sure to be back-compatible.

Does anyone still want to think before merging?