@eric-wieser 2017-11-20
@ahaldane: Happy enough with this for it to make 1.14?